==> notebooks/07_xgboost_model.ipynb <==
import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2024, 6, 1, 0, 0, 0),
    target_column_name='target_rides_next_hour'
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

import xgboost as xgb

# use only past rides data
past_rides_columns = [c for c in X_train.columns if c.startswith('rides_')]
X_train_only_numeric = X_train[past_rides_columns]

from sklearn.base import BaseEstimator
import xgboost as xgb

class CustomXGBRegressor(BaseEstimator):
    """
    A custom wrapper for XGBoost that avoids scikit-learn mixin issues
    """
    def __init__(self, **kwargs):
        self.model = xgb.XGBRegressor(**kwargs)
        self._estimator_type = "regressor"
        
    def fit(self, X, y, **kwargs):
        """Fit the model"""
        self.model.fit(X, y, **kwargs)
        return self
        
    def predict(self, X):
        """Make predictions"""
        return self.model.predict(X)
        
    def get_params(self, deep=True):
        """Get parameters"""
        return self.model.get_params(deep)
        
    def set_params(self, **params):
        """Set parameters"""
        self.model.set_params(**params)
        return self

model = CustomXGBRegressor()
model.fit(X_train_only_numeric, y_train)

X_test_only_numeric = X_test[past_rides_columns]
predictions = model.predict(X_test_only_numeric)
predictions

from sklearn.metrics import mean_absolute_error
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')
